Refactor autospawn.js to export autospawn function
This commit modifies the autospawn.js file to export the autospawn function. Previously, the function was defined but not exported, which prevented it from being used in other modules. By adding the line "module.exports = autospawn;", the autospawn function can now be accessed and used by other parts of the codebase.

Note: This suggested commit message follows the established convention of using the prefix "Refactor" to indicate code changes that improve the structure or organization of the code.