Refactor harvester role to use store.getFreeCapacity()
The code changes in role.harvester.js update the condition for checking if the creep has free capacity to store energy. Instead of using `creep.store[RESOURCE_ENERGY] == 0`, the code now uses `creep.store.getFreeCapacity() > 0`. This change improves the accuracy of determining if the creep can harvest more energy.

Note: This suggested commit message follows the established convention of using the prefix "Refactor" to indicate code changes that improve the structure or organization of the code.