feat: Add role.signer for signing controllers
This commit adds the role.signer module to the codebase. The role.signer module allows creeps to sign controllers in target rooms. This role expands the capabilities of the creeps and enables them to perform the task of signing controllers with a specified message.

Note: This suggested commit message follows the established convention of using the prefix "feat" to indicate the addition of new features or functionalities.