feat: Add role.suppliesSender for sending supplies
This commit adds the role.suppliesSender module to the codebase. The role.suppliesSender module allows creeps to send supplies to other creeps or structures in target rooms. This role expands the capabilities of the creeps and enables them to perform the task of sending supplies efficiently.

Note: This suggested commit message follows the established convention of using the prefix "feat" to indicate the addition of new features or functionalities.