Refactor role.tombraider.js to add collecting mode
This commit refactors the role.tombraider.js file to add a collecting mode. If there are no tombstones in the room, the creep switches to the deposit mode. This change improves the efficiency of the tombraider role by allowing it to adapt its behavior based on the presence of tombstones.

Note: This suggested commit message follows the established convention of using the prefix "Refactor" to indicate code changes that improve the structure or organization of the code.