Refactor autospawn.js to add pathName parameter
This commit refactors the autospawn.js file to add a new parameter, pathName, to the autospawn function. The pathName parameter is added to the memory of the spawned creeps. This change allows for more flexible pathing options for the creeps, improving their navigation capabilities.

Note: This suggested commit message follows the established convention of using the prefix "Refactor" to indicate code changes that improve the structure or organization of the code.