Refactor structure.tower.js to update avoidUsername array
This commit updates the structure.tower.js file to modify the avoidUsername array. The username 'PersonaComune' has been commented out, effectively excluding it from the list of usernames to avoid attacking. This change allows for more flexibility in defining the usernames to avoid, improving the behavior of the tower structure.

Note: This suggested commit message follows the established convention of using the prefix "Refactor" to indicate code changes that improve the structure or organization of the code.