Refactor role.properties.js to add an additional CARRY to the harvester role
This commit modifies the role.properties.js file to add an additional CARRY part to the harvester role. The previous configuration had three CARRY parts, but now it has four. This change allows the harvester creeps to carry more resources, increasing their efficiency in gathering resources.

Note: This suggested commit message follows the established convention of using the prefix "Refactor" to indicate code changes that improve the structure or organization of the code.