Refactor autospawn.js to add labId parameter
This commit modifies the autospawn function in autospawn.js to include a new parameter, labId. This parameter is used to specify the ID of a lab where the creep will transfer resources. By adding this parameter, the code gains flexibility and allows for more targeted resource management.

Note: This suggested commit message follows the established convention of using the prefix "Refactor" to indicate code changes that improve the structure or organization of the code.