feat: Add ciao function to foldertest.js
This commit adds the ciao function to the foldertest.js file. The ciao function logs a greeting message to the console, using the provided name as a parameter. This new function enhances the functionality of the codebase by introducing a new feature.

Note: This suggested commit message follows the established convention of using the prefix "feat" to indicate the addition of a new feature.