Refactor structure.links.js to use transferEnergy method for energy transfer
This commit updates the structure.links.js file to use the transferEnergy method for transferring energy between linked structures. Previously, the code manually transferred energy by updating the energy values of the link objects. Now, the transferEnergy method is called, which handles the energy transfer and returns the result of the operation. This change improves the accuracy and reliability of the energy transfer process.

Note: This suggested commit message follows the established convention of using the prefix "Refactor" to indicate code changes that improve the structure or organization of the code.