feat: Update link association in structure.links.js
This commit updates the link association in the `structure.links.js` file. The `from` value for the first association is changed from `'66ed30939331332979e53f14'` to `'66ed5d56f1409b56d68faeac'`. This modification enhances the functionality of the code by ensuring the correct link association is established.

Note: This suggested commit message follows the established convention of using the prefix "feat" to indicate the addition of a new feature or functionality.